nvim: upgrade plugins
