+++ nvim/lua/plugins/lsp.lua
@@ -1,10 +1,25 @@
+local function lsp_attach(client, bufnr)
+    if client.supports_method "textDocument/documentHighlight" then
+        local group = vim.api.nvim_create_augroup("lsp_document_highlight", { clear = true })
+        vim.api.nvim_create_autocmd({ "CursorHold", "CursorHoldI" }, {
+            group = group,
+            buffer = bufnr,
+            callback = vim.lsp.buf.document_highlight,
+        })
+        vim.api.nvim_create_autocmd("CursorMoved", {
+            group = group,
+            buffer = bufnr,
+            callback = vim.lsp.buf.clear_references,
+        })
+    end
+    vim.lsp.handlers['textDocument/signatureHelp'] = vim.lsp.with(
+        vim.lsp.handlers['signature_help'],
+        { border = 'single', close_events = { "CursorMoved", "BufHidden", "InsertCharPre" } }
+    )
+end
+
 return {
 
-    {
-        "VonHeikemen/lsp-zero.nvim",
-        branch = "v4.x",
-        lazy = true,
-    },
     {
         "hrsh7th/nvim-cmp",
         event = "InsertEnter",
@@ -28,14 +43,18 @@ return {
             -- LSP Support
             { "hrsh7th/cmp-nvim-lsp" },
             { "hrsh7th/cmp-nvim-lsp-signature-help" },
+            -- {
+            --     "mason-org/mason.nvim",
+            --     -- build = function()
+            --     --     pcall(vim.cmd, "MasonUpdate")
+            --     -- end,
+            -- },
             {
-                "williamboman/mason.nvim",
-                version = "1.11.0",
-                build = function()
-                    pcall(vim.cmd, "MasonUpdate")
-                end,
+                "mason-org/mason-lspconfig.nvim",
+                dependencies = {
+                    { "mason-org/mason.nvim" },
+                }
             },
-            { "williamboman/mason-lspconfig.nvim", version = "1.32.0" },
             { "nvimtools/none-ls.nvim" },
 
             -- LSP customizations
@@ -57,7 +76,6 @@ return {
             -- Comment this out for debugging
             vim.lsp.set_log_level("off")
             local inlay = require("plugins.lspconfig.inlay").setup()
-            local lsp = require("plugins.lspconfig.lsp-zero").setup()
             require("mason").setup({})
             require("mason-lspconfig").setup({
                 ensure_installed = {
@@ -74,25 +92,17 @@ return {
                     "svelte",
                     "yamlls",
                     "jsonls",
+                    exclude = {
+                        "rust_analyzer"
+                    }
                 }
             })
-            require("mason-lspconfig").setup_handlers({
-                function(server_name)
-                    -- Autoconfigure the LSP for installed Mason servers
-                    require("lspconfig")[server_name].setup({})
-                end,
-                -- Force no-op for r-a so Meson can't install it
-                rust_analyzer = function()
-                    return true
-                end,
-            })
             require("plugins.lspconfig.fidget")
-            require("plugins.lspconfig.lua").setup({ lsp = lsp })
+            require("plugins.lspconfig.lua").setup()
             require("plugins.lspconfig.typescript").setup({ inlay = inlay })
-            require("plugins.lspconfig.rust").setup({ inlay = inlay, lsp = lsp })
+            require("plugins.lspconfig.rust").setup({ inlay = inlay })
             require("plugins.lspconfig.json").setup()
             require("plugins.lspconfig.null").setup()
-            lsp.setup()
             local nlspsettings = require("nlspsettings")
             nlspsettings.setup({
                 local_settings_dir = ".vim",
@@ -101,6 +111,18 @@ return {
                 append_default_schemas = true,
                 loader = "json",
             })
+            -- Todo: configure LSP zero
+            vim.api.nvim_create_autocmd("LspAttach", {
+                desc = "LSP document highlighting",
+                callback = function(event)
+                    local client = vim.lsp.get_client_by_id(event.data.client_id)
+                    if not client then
+                        return
+                    end
+                    local bufnr = event.buf
+                    lsp_attach(client, bufnr)
+                end
+            });
         end
     },
     {
