+++ nvim/lua/plugins/lspconfig/lua.lua
@@ -1,9 +1,9 @@
 local M = {}
 
 function M.setup(settings)
-    local lspconfig = require("lspconfig")
-    -- (Optional) Configure lua language server for neovim
-    lspconfig.lua_ls.setup(settings["lsp"].nvim_lua_ls())
+    -- vim.lsp.config("lspconfig", {
+    -- -- (Optional) Configure lua language server for neovim
+    -- lspconfig.lua_ls.setup(settings["lsp"].nvim_lua_ls())
 end
 
 return M
