+++ nvim/lua/plugins/lspconfig/typescript.lua
@@ -1,8 +1,7 @@
 local M = {}
 
 function M.setup(_settings)
-    local lspconfig = require("lspconfig")
-    lspconfig.ts_ls.setup({
+    vim.lsp.config("ts_ls", {
         settings = {
             javascript = {
                 inlayHints = {
